requirements.txt
amqp==5.0.6
appdirs==1.4.3
asgiref==3.4.1
attrs==21.2.0
backcall==0.1.0
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
billiard==3.6.4.0
black==21.7b0
celery==5.1.2
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
configparser==5.0.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
Django==3.2.6
django-admin-rangefilter==0.8.1
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-timezone-field==4.2.1
djangorestframework==3.11.2
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
entrypoints==0.3
filelock==3.0.12
httplib2==0.14.0
idna==2.8
inflection==0.5.1
iniconfig==1.1.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.15.2
Jinja2==3.0.1
jsonschema==3.2.0
keyring==18.0.1
kombu==5.1.0
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.0.1
monotonic==1.5
mypy-extensions==0.4.3
numpy==1.20.3
oauthlib==3.1.0
packaging==20.9
parso==0.5.2
pathspec==0.9.0
pbr==5.6.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==11.9.0
platformdirs==2.2.0
pluggy==0.13.1
prompt-toolkit==2.0.10
protobuf==3.6.1
ptyprocess==0.7.0
py==1.10.0
pycparser==2.20
Pygments==2.3.1
PyJWT==2.1.0
pymacaroons==0.13.0
pymemcache==3.5.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.0
pytest==6.2.4
python-crontab==2.5.1
python-dateutil==2.7.3
python-memcached==1.59
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
redis==3.5.3
regex==2021.8.3
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
scipy==1.6.3
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
testresources==2.0.1
toml==0.10.2
tomli==1.2.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.5
wcwidth==0.2.5
zipp==3.5.0
